[css-text] Handling trailing ideographic spaces as regular spaces
The CSS Text 3 specification states on its 'Phase II: Trimming and
Positioning' section:

https://www.w3.org/TR/css-text-3/#white-space-phase-2

3- A sequence at the end of a line (ignoring any intervening inline box
boundaries) of collapsible spaces (U+0020) and/or ideographic spaces
(U+3000) whose white-space value collapses spaces is removed.

4- If there remains any sequence of white space and/or ideographic spaces
(U+3000) at the end of a line: If white-space is set to pre-wrap, the UA
must hang this sequence. It may also visually collapse the character
advance widths of any that would otherwise overflow.

This CL tries to implement such behavior, since we were not considering
this kind of spaces in the hanging/collapsing logic.

Bug: 972952
Change-Id: Idb0aff05f7f423f5b62f5169177c0240917435a7