[LayoutNG] Multicol baseline propagation support.
Spec: https://drafts.csswg.org/css-align/#baseline-export

The "first baseline set" of a multicol container is the baseline of the
first line box in the first column (i,e, ignore all other columns), or
the baseline of the first line box in the first column spanner that has
such a thing, whichever comes first.

Multicol containers have no "last baseline set", and, as such,
fast/multicol/inline-block-baseline.html is invalid, since an
inline-block aligns with the "last baseline set". Remove the test, and
replace it with its antithesis, baseline-008.html in
wpt/css/css-multicol/.

And added a bunch of tests, since this feature seemed to have no
coverage, apart from the deleted test in fast/multicol/.

Bug: 829028
Change-Id: I954de9618fde87dbfa37292786ca6a2dcce9fa1f