[Text Decorations] Compute ink_overflow
Text Decorations were not contributing to ink_overflow,
leading to remants on the screen when invalidated and
causing the mis-sizing of compsited layers containing
the decorations.

This change adds ink overflow calculations for LayoutNG
fragment items. Legacy layout is unaffected.

Tests added include both the paint layer sizing concern
and the invalidation concern.

Bug: 896295
Change-Id: I3b8cfd9d9c4f3efafd380e6718fd2eaf22229f27