WebAuthn: implement credProps extension
WebAuthn L2 adds an extension called credProps that is meant to convey
additional attributes about a WebAuthn credential at registration time
back to the RP. Currently, the only such attribute is an (optional)
boolean 'rk' that indicates whether the registered credential was
client-side discoverable (aka 'resident key'). This is useful in
combination with another L2 feature, the ResidentKeyRequirement enum,
which adds a middle value ("preferred") to the previously boolean
resident key requirement in a WebAuthn registration request.

The 'rk' boolean in the extension output is optional because there are
authenticators that may create a discoverable credential, even though
the browser didn't request one. In particular, the CTAP2.0 authenticator
spec allows for this, but the CTAP2.1 spec doesn't. Hence, if resident
key creation was not requested, Chrome will not set the 'rk' field if
the responding authenticator is a CTAP2.0 device; but it will set 'rk'
to false for CTAP2.1 authenticators.

Bug: 1117630
Change-Id: Ide483fc9d597b3658197646f07e174b85d9984cb