[mathml] Implement special drawing of largeop operator in displaystyle
This implements special drawing of largeop <mo> element in display style
as described in step 3. of [1]. A new algorithm is introduced for that
purpose and in particular only used for <mo> elements that are made of
a single glyph. Existing painting logic for stretchy radical symbol is
reused and adapted to work with largeop operators.

[1] https://mathml-refresh.github.io/mathml-core/#layout-of-operators

Bug: 6606
Change-Id: Icca70f4bf464417809970fecc101e5aa7f149621